Refactor and E2E /users/new :: Go to fix
