+++ E2E/tests/users/new/users.new.spec.js
@@ -20,9 +20,10 @@ test("/users/new :: Go to through /users", async ({ page }) => {
     ]);
 
     // Click on "New User" button
+    const newUserBtn = await page.$('button[href*="users/new"]');
     await Promise.all([
         page.waitForNavigation({ waitUntil: "networkidle0" }),
-        page.click("th > button"),
+        newUserBtn ? newUserBtn.click() : page.click("th > button"),
     ]);
 
     // Check that the page is loaded
