+++ UI/config/axios.js
@@ -1,5 +1,9 @@
 import axios from "axios";
 
+export const axiosBackend = axios.create({
+    baseURL: process.env.SERVER_BACKEND_URL,
+});
+
 export const axiosAPI = axios.create({
     baseURL: "/api",
 });
