--- UI/config/backendAxios.js
@@ -1,6 +1,0 @@
-import axios from "axios";
-
-console.log(process.env.SERVER_BACKEND_URL);
-export const axiosBackend = axios.create({
-    baseURL: process.env.SERVER_BACKEND_URL,
-});
