+++ UI/pages/items/index.jsx
@@ -13,7 +13,7 @@ import Paginate from "../../components/Pagination";
 import thou from "../../utils/thou";
 import { Icon } from "@iconify/react";
 import serverSideRedirectOnUnauth from "../../utils/serverSideRedirectOnUnauth";
-import { axiosBackend } from "../../config/backendAxios";
+import { axiosBackend } from "../../config/axios";
 
 /**
  *
