+++ UI/pages/login.jsx
@@ -3,8 +3,7 @@ import { useRouter } from "next/router";
 import Nav from "../components/Nav";
 import * as yup from "yup";
 import { Formik, Form, Field } from "formik";
-import { axiosAPI } from "../config/axios";
-import { axiosBackend } from "../config/backendAxios";
+import { axiosAPI, axiosBackend } from "../config/axios";
 import { useContext, useEffect } from "react";
 import { AuthContext } from "./_app";
